requirements.txt
asgiref==3.7.1
Django==4.2.1
django-mysql==4.9.0
djangorestframework==3.14.0
mysql-connector-python==8.0.33
django-rest-framework==0.1.0
protobuf==3.20.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
python-dateutil==2.8.2